resolução de conflitos

+++ src/main/java/org/group2/core/model/Professor.java
@@ -79,7 +79,7 @@ public class Professor {
         return dataHoraCadastro;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
